Rename icon glyphs folder
