+++ gui/src/components/Icon.svelte
@@ -44,43 +44,43 @@
 </script>
 
 <script lang="ts">
-  import AddGlyph from './mono/Add.svelte';
-  import BackGlyph from './mono/Back.svelte';
-  import CheckGlyph from './mono/Check.svelte';
-  import DeleteGlyph from './mono/Delete.svelte';
-  import DetailsGlyph from './mono/Details.svelte';
-  import DocGlyph from './mono/Doc.svelte';
-  import EllipsisGlyph from './mono/Ellipsis.svelte';
-  import FeedbackGlyph from './mono/Feedback.svelte';
-  import HomeGlyph from './mono/Home.svelte';
-  import LayersGlyph from './mono/Layers.svelte';
-  import LeftUpGlyph from './mono/LeftUp.svelte';
-  import LogoDenoGlyph from './mono/LogoDeno.svelte';
-  import LogoDiscordGlyph from './mono/LogoDiscord.svelte';
-  import LogoDockerGlyph from './mono/LogoDocker.svelte';
-  import LogoExpressGlyph from './mono/LogoExpress.svelte';
-  import LogoFlaskGlyph from './mono/LogoFlask.svelte';
-  import LogoGhostGlyph from './mono/LogoGhost.svelte';
-  import LogoHasuraGlyph from './mono/LogoHasura.svelte';
-  import LogoLaravelGlyph from './mono/LogoLaravel.svelte';
-  import LogoNextGlyph from './mono/LogoNext.svelte';
-  import LogoNodeGlyph from './mono/LogoNode.svelte';
-  import LogoNuxtGlyph from './mono/LogoNuxt.svelte';
-  import LogoRubyGlyph from './mono/LogoRuby.svelte';
-  import LogoRustGlyph from './mono/LogoRust.svelte';
-  import LogoSlackGlyph from './mono/LogoSlack.svelte';
-  import LogoStrapiGlyph from './mono/LogoStrapi.svelte';
-  import LogoSvelteGlyph from './mono/LogoSvelte.svelte';
-  import LogoTelegramGlyph from './mono/LogoTelegram.svelte';
-  import LogsGlyph from './mono/Logs.svelte';
-  import NetworkingGlyph from './mono/Networking.svelte';
-  import PauseGlyph from './mono/Pause.svelte';
-  import PlayGlyph from './mono/Play.svelte';
-  import PreferencesGlyph from './mono/Preferences.svelte';
-  import ResetGlyph from './mono/Reset.svelte';
-  import StorageGlyph from './mono/Storage.svelte';
-  import VariableGlyph from './mono/Variable.svelte';
-  import WarningGlyph from './mono/Warning.svelte';
+  import AddGlyph from './icon/glyphs/Add.svelte';
+  import BackGlyph from './icon/glyphs/Back.svelte';
+  import CheckGlyph from './icon/glyphs/Check.svelte';
+  import DeleteGlyph from './icon/glyphs/Delete.svelte';
+  import DetailsGlyph from './icon/glyphs/Details.svelte';
+  import DocGlyph from './icon/glyphs/Doc.svelte';
+  import EllipsisGlyph from './icon/glyphs/Ellipsis.svelte';
+  import FeedbackGlyph from './icon/glyphs/Feedback.svelte';
+  import HomeGlyph from './icon/glyphs/Home.svelte';
+  import LayersGlyph from './icon/glyphs/Layers.svelte';
+  import LeftUpGlyph from './icon/glyphs/LeftUp.svelte';
+  import LogoDenoGlyph from './icon/glyphs/LogoDeno.svelte';
+  import LogoDiscordGlyph from './icon/glyphs/LogoDiscord.svelte';
+  import LogoDockerGlyph from './icon/glyphs/LogoDocker.svelte';
+  import LogoExpressGlyph from './icon/glyphs/LogoExpress.svelte';
+  import LogoFlaskGlyph from './icon/glyphs/LogoFlask.svelte';
+  import LogoGhostGlyph from './icon/glyphs/LogoGhost.svelte';
+  import LogoHasuraGlyph from './icon/glyphs/LogoHasura.svelte';
+  import LogoLaravelGlyph from './icon/glyphs/LogoLaravel.svelte';
+  import LogoNextGlyph from './icon/glyphs/LogoNext.svelte';
+  import LogoNodeGlyph from './icon/glyphs/LogoNode.svelte';
+  import LogoNuxtGlyph from './icon/glyphs/LogoNuxt.svelte';
+  import LogoRubyGlyph from './icon/glyphs/LogoRuby.svelte';
+  import LogoRustGlyph from './icon/glyphs/LogoRust.svelte';
+  import LogoSlackGlyph from './icon/glyphs/LogoSlack.svelte';
+  import LogoStrapiGlyph from './icon/glyphs/LogoStrapi.svelte';
+  import LogoSvelteGlyph from './icon/glyphs/LogoSvelte.svelte';
+  import LogoTelegramGlyph from './icon/glyphs/LogoTelegram.svelte';
+  import LogsGlyph from './icon/glyphs/Logs.svelte';
+  import NetworkingGlyph from './icon/glyphs/Networking.svelte';
+  import PauseGlyph from './icon/glyphs/Pause.svelte';
+  import PlayGlyph from './icon/glyphs/Play.svelte';
+  import PreferencesGlyph from './icon/glyphs/Preferences.svelte';
+  import ResetGlyph from './icon/glyphs/Reset.svelte';
+  import StorageGlyph from './icon/glyphs/Storage.svelte';
+  import VariableGlyph from './icon/glyphs/Variable.svelte';
+  import WarningGlyph from './icon/glyphs/Warning.svelte';
 
   export let glyph: IconGlyph;
 </script>
