+++ gui/src/components/icon/glyphs/Add.svelte
