+++ gui/src/components/icon/glyphs/Back.svelte
