+++ gui/src/components/icon/glyphs/Check.svelte
