+++ gui/src/components/icon/glyphs/Delete.svelte
