+++ gui/src/components/icon/glyphs/Details.svelte
