+++ gui/src/components/icon/glyphs/Doc.svelte
