+++ gui/src/components/icon/glyphs/Ellipsis.svelte
