+++ gui/src/components/icon/glyphs/Feedback.svelte
