+++ gui/src/components/icon/glyphs/Home.svelte
