+++ gui/src/components/icon/glyphs/Layers.svelte
