+++ gui/src/components/icon/glyphs/LeftUp.svelte
