+++ gui/src/components/icon/glyphs/LogoDeno.svelte
