+++ gui/src/components/icon/glyphs/LogoDiscord.svelte
