+++ gui/src/components/icon/glyphs/LogoDocker.svelte
