+++ gui/src/components/icon/glyphs/LogoFlask.svelte
