+++ gui/src/components/icon/glyphs/LogoGhost.svelte
