+++ gui/src/components/icon/glyphs/LogoHasura.svelte
