+++ gui/src/components/icon/glyphs/LogoLaravel.svelte
