+++ gui/src/components/icon/glyphs/LogoNext.svelte
