+++ gui/src/components/icon/glyphs/LogoNode.svelte
