+++ gui/src/components/icon/glyphs/LogoNuxt.svelte
