+++ gui/src/components/icon/glyphs/LogoRuby.svelte
