+++ gui/src/components/icon/glyphs/LogoRust.svelte
