+++ gui/src/components/icon/glyphs/LogoSlack.svelte
