+++ gui/src/components/icon/glyphs/LogoStrapi.svelte
