+++ gui/src/components/icon/glyphs/LogoSvelte.svelte
