+++ gui/src/components/icon/glyphs/Logs.svelte
