+++ gui/src/components/icon/glyphs/Networking.svelte
