+++ gui/src/components/icon/glyphs/Pause.svelte
