+++ gui/src/components/icon/glyphs/Play.svelte
