+++ gui/src/components/icon/glyphs/Preferences.svelte
