+++ gui/src/components/icon/glyphs/Reset.svelte
