+++ gui/src/components/icon/glyphs/Storage.svelte
