+++ gui/src/components/icon/glyphs/Variable.svelte
