+++ gui/src/components/icon/glyphs/Warning.svelte
